requirements.txt
Pillow==9.4.0
Pillow==10.0.1
streamlit==1.26.0
